Win64 warning squashing (Alexey Melnikov <Alexey.Melnikov@isode.com>)

+++ usr/src/lib/libsasl/lib/common.c
@@ -6,7 +6,7 @@
 /* common.c - Functions that are common to server and clinet
  * Rob Siemborski
  * Tim Martin
- * $Id: common.c,v 1.96 2003/09/02 15:34:02 rjs3 Exp $
+ * $Id: common.c,v 1.97 2003/10/20 15:19:58 rjs3 Exp $
  */
 /* 
  * Copyright (c) 1998-2003 Carnegie Mellon University.  All rights reserved.
@@ -212,7 +212,7 @@ int _sasl_strdup(const char *in, char **out, size_t *outlen)
 {
   size_t len = strlen(in);
   if (outlen) *outlen = len;
-  *out=sasl_ALLOC(len + 1);
+  *out=sasl_ALLOC((unsigned) len + 1);
   if (! *out) return SASL_NOMEM;
   strcpy((char *) *out, in);
   return SASL_OK;
@@ -336,7 +336,7 @@ int sasl_encodev(sasl_conn_t *conn,
 	if(result != SASL_OK) INTERROR(conn, result);
        
 	*output = conn->encode_buf->data;
-	*outputlen = conn->encode_buf->curlen;
+	*outputlen = (unsigned) conn->encode_buf->curlen;
 
 #ifdef _INTEGRATED_SOLARIS_
     } else if (!conn->sun_reg) {
@@ -992,12 +992,12 @@ int sasl_setprop(sasl_conn_t *conn, int propnum, const void *value)
 	      ((sasl_client_conn_t *)conn)->cparams->ipremoteport
 		  = conn->ipremoteport;
 	      ((sasl_client_conn_t *)conn)->cparams->ipremlen =
-		  strlen(conn->ipremoteport);
+		  (unsigned) strlen(conn->ipremoteport);
 	  } else if (conn->type == SASL_CONN_SERVER) {
 	      ((sasl_server_conn_t *)conn)->sparams->ipremoteport
 		  = conn->ipremoteport;
 	      ((sasl_server_conn_t *)conn)->sparams->ipremlen =
-		  strlen(conn->ipremoteport);
+		  (unsigned) strlen(conn->ipremoteport);
 	  }
       } else {
 	  if(conn->type == SASL_CONN_CLIENT) {
@@ -1041,12 +1041,12 @@ int sasl_setprop(sasl_conn_t *conn, int propnum, const void *value)
 	      ((sasl_client_conn_t *)conn)->cparams->iplocalport
 		  = conn->iplocalport;
 	      ((sasl_client_conn_t *)conn)->cparams->iploclen
-		  = strlen(conn->iplocalport);
+		  = (unsigned) strlen(conn->iplocalport);
 	  } else if (conn->type == SASL_CONN_SERVER) {
 	      ((sasl_server_conn_t *)conn)->sparams->iplocalport
 		  = conn->iplocalport;
 	      ((sasl_server_conn_t *)conn)->sparams->iploclen
-		  = strlen(conn->iplocalport);
+		  = (unsigned) strlen(conn->iplocalport);
 	  }
       } else {
 	  if(conn->type == SASL_CONN_CLIENT) {
@@ -1269,7 +1269,7 @@ const char *sasl_errdetail(sasl_conn_t *conn)
     snprintf(leader,128,"SASL(%d): %s: ",
 	     sasl_usererr(conn->error_code), errstr);
     
-    need_len = strlen(leader) + strlen(conn->error_buf) + 12;
+    need_len = (unsigned) (strlen(leader) + strlen(conn->error_buf) + 12);
 #ifdef _SUN_SDK_
     ret = _buf_alloc(&conn->errdetail_buf, &conn->errdetail_buf_len, need_len);
     if (ret != SASL_OK)
@@ -1387,7 +1387,7 @@ static int _sasl_global_getopt(void *context,
   *result = sasl_config_getstring(option, NULL);
 #endif /* _SUN_SDK_ */
   if (*result != NULL) {
-      if (len) { *len = strlen(*result); }
+      if (len) { *len = (unsigned) strlen(*result); }
       return SASL_OK;
   }
 
@@ -2193,7 +2193,7 @@ int _buf_alloc(char **rwbuf, size_t *curlen, size_t newlen)
 #endif /* _SUN_SDK_ */
 {
     if(!(*rwbuf)) {
-	*rwbuf = sasl_ALLOC(newlen);
+	*rwbuf = sasl_ALLOC((unsigned)newlen);
 	if (*rwbuf == NULL) {
 	    *curlen = 0;
 	    return SASL_NOMEM;
@@ -2205,7 +2205,7 @@ int _buf_alloc(char **rwbuf, size_t *curlen, size_t newlen)
 	while(needed < newlen)
 	    needed *= 2;
 
-	*rwbuf = sasl_REALLOC(*rwbuf, needed);
+	*rwbuf = sasl_REALLOC(*rwbuf, (unsigned)needed);
 	
 	if (*rwbuf == NULL) {
 	    *curlen = 0;
